Use chi as http mux
