+++ assets/example-multipleServers-pretty.html
@@ -13,18 +13,18 @@ var SPEEDTEST_SERVERS=[
 	{	//this is my demo server, remove it
 		name:"Speedtest Demo Server (Helsinki)", //user friendly name for the server
 		server:"//fi.openspeed.org/", //URL to the server. // at the beginning will be replaced with http:// or https:// automatically
-		dlURL:"garbage.php",  //path to download test on this server (garbage.php or replacement)
-		ulURL:"empty.php",  //path to upload test on this server (empty.php or replacement)
-		pingURL:"empty.php",  //path to ping/jitter test on this server (empty.php or replacement)
-		getIpURL:"getIP.php"  //path to getIP on this server (getIP.php or replacement)
+		dlURL:"garbage",  //path to download test on this server (garbage.php or replacement)
+		ulURL:"empty",  //path to upload test on this server (empty.php or replacement)
+		pingURL:"empty",  //path to ping/jitter test on this server (empty.php or replacement)
+		getIpURL:"getIP"  //path to getIP on this server (getIP.php or replacement)
 	},
 	{	//this is my demo server, remove it
 		name:"Old Speedtest Demo Server",
 		server:"//mpotdemo.fdossena.com/",
-		dlURL:"garbage.php",
-		ulURL:"empty.php",
-		pingURL:"empty.php",
-		getIpURL:"getIP.php"
+		dlURL:"garbage",
+		ulURL:"empty",
+		pingURL:"empty",
+		getIpURL:"getIP"
 	}
 	//add other servers here, comma separated
 ];
