+++ assets/speedtest.js
@@ -20,10 +20,10 @@
         {
             name: "User friendly name",
             server:"http://yourBackend.com/",     <---- URL to your server. You can specify http:// or https://. If your server supports both, just write // without the protocol
-            dlURL:"garbage.php"    <----- path to garbage.php or its replacement on the server
-            ulURL:"empty.php"    <----- path to empty.php or its replacement on the server
-            pingURL:"empty.php"    <----- path to empty.php or its replacement on the server. This is used to ping the server by this selector
-            getIpURL:"getIP.php"    <----- path to getIP.php or its replacement on the server
+            dlURL:"garbage"    <----- path to garbage.php or its replacement on the server
+            ulURL:"empty"    <----- path to empty.php or its replacement on the server
+            pingURL:"empty"    <----- path to empty.php or its replacement on the server. This is used to ping the server by this selector
+            getIpURL:"getIP"    <----- path to getIP.php or its replacement on the server
         }
         While in state 1, you can only add test points, you cannot change the test settings. When you're done, use selectServer(callback) to select the test point with the lowest ping. This is asynchronous, when it's done, it will call your callback function and move to state 2. Calling setSelectedServer(server) will manually select a server and move to state 2.
     - 2: test point selected, ready to start the test. Use start() to begin, this will move to state 3
