+++ assets/speedtest_worker.js
@@ -11,7 +11,7 @@ var dlStatus = ""; // download speed in megabit/s with 2 decimal digits
 var ulStatus = ""; // upload speed in megabit/s with 2 decimal digits
 var pingStatus = ""; // ping in milliseconds with 2 decimal digits
 var jitterStatus = ""; // jitter in milliseconds with 2 decimal digits
-var clientIp = ""; // client's IP address as reported by getIP.php
+var clientIp = ""; // client's IP address as reported by getIP
 var dlProgress = 0; //progress of download test 0-1
 var ulProgress = 0; //progress of upload test 0-1
 var pingProgress = 0; //progress of ping+jitter test 0-1
@@ -45,10 +45,10 @@ var settings = {
 	time_ulGraceTime: 3, //time to wait in seconds before actually measuring ul speed (wait for buffers to fill)
 	time_dlGraceTime: 1.5, //time to wait in seconds before actually measuring dl speed (wait for TCP window to increase)
 	count_ping: 10, // number of pings to perform in ping test
-	url_dl: "backend/garbage.php", // path to a large file or garbage.php, used for download test. must be relative to this js file
-	url_ul: "backend/empty.php", // path to an empty file, used for upload test. must be relative to this js file
-	url_ping: "backend/empty.php", // path to an empty file, used for ping test. must be relative to this js file
-	url_getIp: "backend/getIP.php", // path to getIP.php relative to this js file, or a similar thing that outputs the client's ip
+	url_dl: "backend/garbage", // path to a large file or garbage.php, used for download test. must be relative to this js file
+	url_ul: "backend/empty", // path to an empty file, used for upload test. must be relative to this js file
+	url_ping: "backend/empty", // path to an empty file, used for ping test. must be relative to this js file
+	url_getIp: "backend/getIP", // path to getIP.php relative to this js file, or a similar thing that outputs the client's ip
 	getIp_ispInfo: true, //if set to true, the server will include ISP info with the IP address
 	getIp_ispInfo_distance: "km", //km or mi=estimate distance from server in km/mi; set to false to disable distance estimation. getIp_ispInfo must be enabled in order for this to work
 	xhr_dlMultistream: 6, // number of download streams to use (can be different if enable_quirks is active)
@@ -63,7 +63,7 @@ var settings = {
 	overheadCompensationFactor: 1.06, //can be changed to compensatie for transport overhead. (see doc.md for some other values)
 	useMebibits: false, //if set to true, speed will be reported in mebibits/s instead of megabits/s
 	telemetry_level: 0, // 0=disabled, 1=basic (results only), 2=full (results and timing) 3=debug (results+log)
-	url_telemetry: "results/telemetry.php", // path to the script that adds telemetry data to the database
+	url_telemetry: "results/telemetry", // path to the script that adds telemetry data to the database
 	telemetry_extra: "" //extra data that can be passed to the telemetry through the settings
 };
