+++ database/database.go
@@ -3,18 +3,13 @@ package database
 import (
 	"database/sql"
 
+	"backend/config"
 	"backend/database/mysql"
 	"backend/database/postgresql"
 	"backend/database/schema"
 )
 
 var (
-	Type     string
-	Hostname string
-	Username string
-	Password string
-	Database string
-
 	DB DataAccess
 )
 
@@ -23,26 +18,11 @@ type DataAccess interface {
 	FetchByUUID(string) (*schema.TelemetryData, error)
 }
 
-func SetDBInfo(settings map[string]string) {
-	for k, v := range settings {
-		switch k {
-		case "database_type":
-			Type = v
-		case "database_hostname":
-			Hostname = v
-		case "database_name":
-			Database = v
-		case "database_username":
-			Username = v
-		case "database_password":
-			Password = v
-		}
-	}
-
-	switch Type {
+func SetDBInfo(conf *config.Config) {
+	switch conf.DatabaseType {
 	case "postgresql":
-		DB = postgresql.Open(Hostname, Username, Password, Database)
+		DB = postgresql.Open(conf.DatabaseHostname, conf.DatabaseUsername, conf.DatabasePassword, conf.DatabaseName)
 	case "mysql":
-		DB = mysql.Open(Hostname, Username, Password, Database)
+		DB = mysql.Open(conf.DatabaseHostname, conf.DatabaseUsername, conf.DatabasePassword, conf.DatabaseName)
 	}
 }
