+++ go.mod
@@ -3,9 +3,15 @@ module backend
 go 1.13
 
 require (
+	github.com/go-chi/chi v4.0.3+incompatible
+	github.com/go-chi/cors v1.0.0
+	github.com/go-chi/render v1.0.1
 	github.com/go-sql-driver/mysql v1.5.0
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
 	github.com/lib/pq v1.3.0
 	github.com/oklog/ulid/v2 v2.0.2
+	github.com/sirupsen/logrus v1.4.2
+	github.com/spf13/viper v1.6.2
 	golang.org/x/image v0.0.0-20200119044424-58c23975cae1
+	gopkg.in/yaml.v2 v2.2.4
 )
